requirements.txt
streamlit
streamlit-webrtc
opencv-python
mediapipe
tensorflow
numpy
pillow

